Set working directory actively for every hook execution.
Every hook execution will be executed in the root directory of the configured repository.

Variable of plugins can be interpolated with giks variables. TODO: Enhance syntax and mechanism to work with env vars and commands.

Introduce list-comparator plugin to compare lists against each other and fail in case of a match.